+++ src/util-file.h
@@ -110,7 +110,7 @@ void FileContainerAdd(FileContainer *, File *);
  *  \note filename is not a string, so it's not nul terminated.
  */
 File *FileOpenFile(FileContainer *, const uint8_t *name, uint16_t name_len,
-        const uint8_t *data, uint32_t data_len, uint8_t flags);
+        const uint8_t *data, uint32_t data_len, uint16_t flags);
 /**
  *  \brief Close a File
  *
@@ -123,7 +123,7 @@ File *FileOpenFile(FileContainer *, const uint8_t *name, uint16_t name_len,
  *  \retval -1 error
  */
 int FileCloseFile(FileContainer *, const uint8_t *data, uint32_t data_len,
-        uint8_t flags);
+        uint16_t flags);
 
 /**
  *  \brief Store a chunk of file data in the flow. The open "flowfile"
